Fix findings from full-stack code review (2026-09-15)
P0: store money as Decimal/NUMERIC(12,2) instead of float end-to-end
(Transaction.amount, Account.opening_balance, Budget.limit_amount/
spent_amount/remaining_amount) so SUM()s and balance math can't drift
the way binary floats do. Response schemas use a new Money type that
still serializes to a plain JSON number, so the wire contract — and
every frontend consumer — is unchanged. Existing Postgres deploys get
migrated via an ALTER ... USING cast in migrate_schema().

P1 security hardening:
- Login and forgot-password no longer leak whether an identifier is
  registered (unified 401 / always-200 responses).
- Rate limiting on /login and /forgot-password (slowapi, per-IP).
- Gmail-connect and Google sign-in no longer put bearer tokens in
  URLs — both now use short-lived, single-use exchange tokens instead
  of the raw access token / JWT.

P2:
- users.phone now has a real DB-level unique index, not just an
  app-level check (closes a concurrent-signup race).
- Gmail/SMS imports enforce the same amount ceiling manual entry gets.
- Manual transaction creation supports an idempotency key (DB unique
  index) so a retried create can't double-insert.
- Gmail token revoke failures are logged instead of silently
  swallowed.

Plus a few nits (dead datetime.utcnow() in a test, unused test vars).

98/98 tests passing.

+++ app/models/account.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Boolean, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, Boolean, DateTime, ForeignKey
 from app.config.database import Base
 from app.utils.timezone import now_ist
 
@@ -14,7 +14,8 @@ class Account(Base):
     name = Column(String, nullable=False)
     type = Column(String, nullable=False)  # bank, cash, credit_card, wallet, investment
     currency = Column(String, default="INR")
-    opening_balance = Column(Float, nullable=False, default=0)
+    # NUMERIC, not FLOAT — see the same note on Transaction.amount.
+    opening_balance = Column(Numeric(12, 2), nullable=False, default=0)
     is_archived = Column(Boolean, default=False)
     created_at = Column(DateTime, default=now_ist)
     updated_at = Column(DateTime, default=now_ist, onupdate=now_ist)
