+++ app/models/budget.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Computed, Integer, String, Float, DateTime, ForeignKey
+from sqlalchemy import Column, Computed, Integer, String, Numeric, DateTime, ForeignKey
 from app.config.database import Base
 from app.utils.timezone import now_ist
 
@@ -8,12 +8,13 @@ class Budget(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
     category = Column(String, nullable=False)
-    limit_amount = Column(Float, nullable=False)
-    spent_amount = Column(Float, default=0)
+    # NUMERIC, not FLOAT — see the same note on Transaction.amount.
+    limit_amount = Column(Numeric(12, 2), nullable=False)
+    spent_amount = Column(Numeric(12, 2), default=0)
     # Always limit_amount - spent_amount. A DB-generated column rather than a
     # third value the app has to keep in sync: the database recomputes it on
     # every write and it can never drift. Read-only from the ORM's side.
-    remaining_amount = Column(Float, Computed("limit_amount - spent_amount", persisted=True))
+    remaining_amount = Column(Numeric(12, 2), Computed("limit_amount - spent_amount", persisted=True))
     period = Column(String, nullable=False)  # weekly, monthly
     start_date = Column(DateTime, nullable=False)
     end_date = Column(DateTime, nullable=False)
